+++ pennylane_lightning/core/bindings/Bindings_nb.cpp
@@ -27,6 +27,7 @@
 namespace Pennylane::NanoBindings {
 
 #if defined(LIGHTNING_MODULE_NAME)
+
 /**
  * @brief Add Lightning State-vector C++ classes, methods and functions to
  * Python module using Nanobind.
@@ -57,10 +58,13 @@ NB_MODULE(LIGHTNING_MODULE_NAME, m) {
  * module.
  */
 NB_MODULE(LIGHTNING_TENSOR_MODULE_NAME, m) {
+    // Register general info
+    registerInfo(m);
+
     // Register bindings for backend-specific info:
     registerBackendSpecificInfo(m);
 
-    registerLightningTensorClassBindings<TensorNetworkBackends>(m);
+    registerLightningClassBindings<TensorNetworkBackends>(m);
 }
 #endif
 } // namespace Pennylane::NanoBindings